bt-224: * rename step, delete comments
