+++ frontend/src/bundles/talent-onboarding/components/badges-step/constants/constants.ts
