+++ frontend/src/bundles/talent-onboarding/components/badges-step/constants/default.constants.ts
@@ -1,7 +1,7 @@
 import { type BsaBadgesStepDto } from '~/bundles/talent-onboarding/types/types.js';
 
 const DEFAULT_PAYLOAD_BSA_BADGES_STEP: BsaBadgesStepDto = {
-    bsaBadges: [],
+    badges: [],
 };
 
 export { DEFAULT_PAYLOAD_BSA_BADGES_STEP };
