+++ frontend/src/bundles/talent-onboarding/components/badges-step/styles.module.scss
