+++ frontend/src/bundles/talent-onboarding/components/components.tsx
@@ -1,4 +1,4 @@
-export { BsaBadgesStep } from './bsa-badges-step/bsa-badges-step.js';
+export { BadgesStep } from './badges-step/badges-step.js';
 export { ContactsCVStep } from './contacts-cv-step/contacts-cv-step.js';
 export { ProfileStep } from './profile-step/profile-step.js';
 export { SkillsStep } from './skills-step/skills-step.js';
