+++ frontend/src/bundles/talent-onboarding/store/slice.ts
@@ -1,7 +1,7 @@
 import { createSlice } from '@reduxjs/toolkit';
 import { type UserDetailsUpdateRequestDto } from 'shared/build/index.js';
 
-import { DEFAULT_PAYLOAD_BSA_BADGES_STEP } from '../components/bsa-badges-step/constants/constants.js';
+import { DEFAULT_PAYLOAD_BSA_BADGES_STEP } from '../components/badges-step/constants/constants.js';
 import { DEFAULT_CONTACTS_CV_STEP_PAYLOAD } from '../components/contacts-cv-step/constants/constants.js';
 import { DEFAULT_PAYLOAD_PROFILE_STEP } from '../components/profile-step/constants/default.constants.js';
 import { DEFAULT_PAYLOAD_SKILLS_STEP } from '../components/skills-step/constants/default.constants.js';
@@ -13,7 +13,7 @@ import { updateTalentDetails } from './actions.js';
 const initialState: UserDetailsGeneralCustom = {
     ...DEFAULT_PAYLOAD_PROFILE_STEP,
     ...DEFAULT_PAYLOAD_BSA_BADGES_STEP,
-    bsaBadges: mockBadges
+    badges: mockBadges
         .filter((badge) => badge.type === 'service')
         .map((badge) => badge.id),
     ...DEFAULT_PAYLOAD_SKILLS_STEP,
