+++ frontend/src/bundles/talent-onboarding/types/general-update-user-details-request-dto/general-update-user-details-request-dto.ts
@@ -5,7 +5,7 @@ type UserDetailsGeneralCustom = UserDetailsUpdateRequestDto & {
         value: string;
         label: string;
     }[];
-    bsaBadges?: string[];
+    badges?: string[];
     photo?: File | null;
     cv?: File | null;
 };
