+++ frontend/src/bundles/talent-onboarding/types/types.ts
@@ -1,5 +1,5 @@
 export { type ContactsCVStepDto } from './contacts-cv-step/contacts-cv-step-dto.js';
-export { type UserDetailsGeneralCustom } from './update-user-details-request-dto/general-update-user-details-request-dto.js';
+export { type UserDetailsGeneralCustom } from './general-update-user-details-request-dto/general-update-user-details-request-dto.js';
 export { type BsaBadgesStepDto } from 'shared/build/index.js';
 export { type ProfileStepDto } from 'shared/build/index.js';
 export { type SkillsStepDto } from 'shared/build/index.js';
