+++ shared/src/bundles/talent-onboarding/types/bsa-badges-step/bsa-badges-step-dto.ts
@@ -1,5 +1,5 @@
 type BsaBadgesStepDto = {
-    bsaBadges: string[];
+    badges: string[];
 };
 
 export { type BsaBadgesStepDto };
